Add files via upload
This is an Android Application that connects to the server side through socket. It Reads up your mobile accelerometer readings in real time and passes it to the server side in real time.
For connecting to the server side the IP Address of the Server should be added and the port number in which the Server is listening should be entered and then you can connect to the server through socket and send the real time accelerometer reading to the server.